+++ test/blackbox-tests/test-cases/rule-target-inferrence/run.t
@@ -9,8 +9,4 @@ But should work with the long form as well:
 
   $ dune build @infer --root long-form
   Entering directory 'long-form'
-  File "dune", line 1, characters 0-57:
-  1 | (rule
-  2 |  (action (write-file b "It should work as well!")))
-  Error: fields targets, target are all missing (exactly one is needed)
-  [1]
+  It should work as well!